rename test_data -> validation data
The `FitResult` object uses `test_losses` as a key in the output. For
compatibility I did not modify this field for now. But the `fit` method
uses `validation_data` in the docs.

Need to confirm whether or not you can still load a `FitResult` after
the name change of variables or not. If not then we should update the
field now.