fix: scoring is close to tolerance now
There are still slight differences in the tests that cause failures. I
think these are due to the way vectorization is handled at the C++ level
and there probably isn't much we can do to improve that.

The benefit of this refactor is that we can amortize the cost of scoring
across multiple fields at once instead of doing all of the work once per
field that we score.