relabel test loss in plotting and other interfaces
should we rename test_loss in FitResult and other interface objects?
Note it's a breaking change for serialization/deserialization, but we
are already making breaking changes in this branch elsewhere by renaming
things to `validation_data`. If we are going to make those changes at
all then we should do them all at once in one place.